backend/requirements.txt
fastapi
uvicorn
scikit-learn
pandas
sqlalchemy
joblib
python-multipart
pydantic
PyGithub
httpx
